Rename to LazyCodeGenerator
