+++ e3nn/util/__init__.py
@@ -1,2 +1,1 @@
-from .eval_code import *  # noqa
 from .default_type import *  # noqa
