+++ e3nn/util/codegen/__init__.py
@@ -1,0 +1,3 @@
+from ._eval import eval_code  # noqa
+from ._mixin import CodeGenMixin  # noqa
+from ._lazy import LazyCodeGenerator  # noqa
